+++ activerecord/lib/active_record/connection_adapters/abstract_adapter.rb
@@ -279,6 +279,11 @@ module ActiveRecord
         false
       end
 
+      # Does this adapter support virtual columns?
+      def supports_virtual_columns?
+        false
+      end
+
       # This is meant to be implemented by the adapters that support extensions
       def disable_extension(name)
       end