+++ activerecord/lib/active_record/connection_adapters/mysql/column.rb
@@ -31,6 +31,10 @@ module ActiveRecord
           extra == 'auto_increment'
         end
 
+        def virtual?
+          /(?:VIRTUAL|STORED) GENERATED/ === extra
+        end
+
         private
 
         def extract_default